refactor dependency utils
