+++ graphein/utils/utils.py
@@ -333,49 +333,6 @@ def protein_letters_3to1_all_caps(amino_acid: str) -> str:
     return protein_letters_3to1[amino_acid]
 
 
-def import_message(
-    submodule: str,
-    package: str,
-    conda_channel: Optional[str] = None,
-    pip_install: bool = False,
-) -> str:
-    """
-    Return warning if package is not found.
-    Generic message for indicating to the user when a function relies on an
-    optional module / package that is not currently installed. Includes
-    installation instructions. Typically used in conjunction without optional featurisation libraries
-
-    :param submodule: graphein submodule that needs an external dependency.
-    :type submodule: str
-    :param package: External package this submodule relies on.
-    :type package: str
-    :param conda_channel: Conda channel package can be installed from, if at all. Defaults to ``None``.
-    :type conda_channel: str, optional
-    :param pip_install: Whether package can be installed via pip. Defaults to ``False``.
-    :type pip_install: bool
-    """
-    is_conda = os.path.exists(os.path.join(sys.prefix, "conda-meta"))
-    installable = True
-    if is_conda:
-        if conda_channel is None:
-            installable = False
-            installation = f"{package} cannot be installed via conda"
-        else:
-            installation = f"conda install -c {conda_channel} {package}"
-    elif pip_install:
-        installation = f"pip install {package}"
-    else:
-        installable = False
-        installation = f"{package} cannot be installed via pip"
-
-    message = f"To use the Graphein submodule {submodule}, you need to install: {package} "
-    if installable:
-        message += f"\nTo do so, use the following command: {installation}"
-    else:
-        message += f"\n{installation}"
-    return message
-
-
 def ping(host: str) -> bool:
     """
     Returns ``True`` if host (str) responds to a ping request.
